requirements.txt
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
lxml==5.4.0
mypy_extensions==1.1.0
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyfiglet==1.0.2
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python3-discogs-client==2.8
pyzbar==0.1.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tabulate==0.9.0
termcolor==3.1.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
